requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-tenants>=3.5.0
celery>=5.3.0
redis>=4.6.0
Pillow>=10.0.0
python-dotenv>=1.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.0
